moved all commands to commands subfolder. Fixed a bug where several command subclases were overriding xecute method instead of doExecute.
