+++ src/commands/BestExploitCommand.cc
@@ -4,7 +4,7 @@ BestExploitCommand::BestExploitCommand() :
 		Command() {
 }
 
-int BestExploitCommand::execute() {
+int BestExploitCommand::doExecute() {
 	if (input_token.size() < 3) {
 		Display->Warning("n. of iterations and a phrase should be provided");
 		Display->Warning("as argument in the current version.");
