+++ src/commands/BestExploitCommand.h
@@ -9,7 +9,7 @@
 class BestExploitCommand: public Command {
 public:
 	BestExploitCommand();
-	int execute();
+	int doExecute();
 };
 
 #endif /* SRC_BESTEXPLOITCOMMAND_H_ */
