+++ src/commands/CleanExploitCommand.cc
@@ -1,10 +1,13 @@
-#include "ExploitCommand.h"
+#include "CleanExploitCommand.h"
 
-ExploitCommand::ExploitCommand() :
+CleanExploitCommand::CleanExploitCommand() :
 		Command() {
 }
 
-int ExploitCommand::execute() {
+int CleanExploitCommand::doExecute() {
+	for (int i = 0; i < 5; i++)
+		ExecuteAct(annabell, Mon, STORE_ST_A, NULL_ACT, DROP_GL);
+
 	if (input_token.size() > 1) {
 		string target_phrase;
 		target_phrase = input_token[1];
@@ -25,4 +28,5 @@ int ExploitCommand::execute() {
 	CheckSensoryMotor(annabell->flags->OutPhrase, annabell, Display);
 
 	return 0;
+
 }
