+++ src/commands/CleanExploitCommand.h
@@ -13,7 +13,7 @@
 class CleanExploitCommand: public Command {
 public:
 	CleanExploitCommand();
-	int execute();
+	int doExecute();
 };
 
 #endif /* SRC_CLEANEXPLOITCOMMAND_H_ */
