+++ src/commands/ContinueContextCommand.cc
