+++ src/commands/ContinueContextCommand.h
