+++ src/commands/ContinueSearchPhraseCommand.cc
