+++ src/commands/ContinueSearchPhraseCommand.h
