+++ src/commands/CopyInputPhraseToWorkingPhraseCommand.cc
