+++ src/commands/ExploitCommand.h
@@ -13,7 +13,7 @@
 class ExploitCommand: public Command {
 public:
 	ExploitCommand();
-	int execute();
+	int doExecute();
 };
 
 #endif /* SRC_EXPLOITCOMMAND_H_ */
