+++ src/commands/ExploitMemorizeCommand.cc
@@ -1,10 +1,10 @@
-#include "ExploitRandomCommand.h"
+#include "ExploitMemorizeCommand.h"
 
-ExploitRandomCommand::ExploitRandomCommand() :
+ExploitMemorizeCommand::ExploitMemorizeCommand() :
 		Command() {
 }
 
-int ExploitRandomCommand::execute() {
+int ExploitMemorizeCommand::doExecute() {
 	if (input_token.size() > 1) {
 		string target_phrase;
 		target_phrase = input_token[1];
@@ -14,13 +14,14 @@ int ExploitRandomCommand::execute() {
 		ExecuteAct(annabell, Mon, STORE_ST_A, NULL_ACT, FLUSH_OUT);
 		GetInputPhrase(annabell, Mon, target_phrase);
 	}
-	annabell->RemPhfWG->OrderedWnnFlag = false;
 	//VerboseFlag = true;
 	//string out_phrase=ExploitationTest(annabell, Mon, 1);
 	annabell->flags->OutPhrase = Exploitation(annabell, Mon, Display, 1);
 	annabell->flags->CompleteOutputFlag = true;
 	//VerboseFlag = false;
-	annabell->RemPhfWG->OrderedWnnFlag = true;
+	ExecuteAct(annabell, Mon, STORE_ST_A, NULL_ACT, FLUSH_OUT);
+	GetInputPhrase(annabell, Mon, annabell->flags->OutPhrase);
+	BuildAs(annabell, Mon);
 	// check if the output is a sensorymotor command
 	CheckSensoryMotor(annabell->flags->OutPhrase, annabell, Display);
