+++ src/commands/ExploitMemorizeCommand.h
@@ -9,7 +9,7 @@
 class ExploitMemorizeCommand: public Command {
 public:
 	ExploitMemorizeCommand();
-	int execute();
+	int doExecute();
 };
 
 #endif /* SRC_EXPLOITMEMORIZECOMMAND_H_ */
