+++ src/commands/ExploitRandomCommand.h
@@ -13,7 +13,7 @@
 class ExploitRandomCommand: public Command {
 public:
 	ExploitRandomCommand();
-	int execute();
+	int doExecute();
 };
 
 #endif /* SRC_EXPLOITRANDOMCOMMAND_H_ */
