+++ src/commands/ExtractWordGroupCommand.cc
