+++ src/commands/ExtractWordGroupCommand.h
