+++ src/commands/InputPhraseCommand.cc
