+++ src/commands/QuitCommand.cc
