+++ src/commands/RetrievePhraseCommand.cc
