+++ src/commands/SearchPhraseCommand.cc
