+++ src/yarpinterface.cc
@@ -15,7 +15,7 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "CommandFactory.h"
+#include "commands/CommandFactory.h"
 #include "sensorymotor.h"
 using namespace std;
